+++ benchmark/formatter_bench_test.go
@@ -4,20 +4,19 @@ import (
 	"testing"
 
 	"github.com/xybor-x/xylog"
+	"github.com/xybor-x/xylog/test"
 )
 
 func BenchmarkTextFormatter(b *testing.B) {
-	var record = xylog.LogRecord{}
-	var formatter = addFullMacros(xylog.NewTextFormatter())
+	var formatter = test.AddFullMacros(xylog.NewTextFormatter())
 	for i := 0; i < b.N; i++ {
-		formatter.Format(record)
+		formatter.Format(test.FullRecord)
 	}
 }
 
 func BenchmarkJSONFormatter(b *testing.B) {
-	var record = xylog.LogRecord{}
-	var formatter = addFullMacros(xylog.NewJSONFormatter())
+	var formatter = test.AddFullMacros(xylog.NewJSONFormatter())
 	for i := 0; i < b.N; i++ {
-		formatter.Format(record)
+		formatter.Format(test.FullRecord)
 	}
 }