+++ benchmark/logger_bench_test.go
@@ -4,41 +4,55 @@ import (
 	"testing"
 
 	"github.com/xybor-x/xylog"
+	"github.com/xybor-x/xylog/test"
 )
 
 var DevnullEmitter *xylog.StreamEmitter
 
-func BenchmarkLoggerWithoutLog(b *testing.B) {
-	withBenchLogger(b, func(logger *xylog.Logger) {
+func BenchmarkLoggerDisable(b *testing.B) {
+	test.WithBenchLogger(b, func(logger *xylog.Logger) {
+		logger.SetLevel(xylog.ERROR)
 		b.RunParallel(func(p *testing.PB) {
 			for p.Next() {
-				logger.Debug(getRandomMessage())
+				logger.Debug(test.GetRandomMessage())
+			}
+		})
+	})
+}
+
+func BenchmarkLoggerWithoutHandler(b *testing.B) {
+	test.WithBenchLogger(b, func(logger *xylog.Logger) {
+		logger.RemoveHandler(logger.Handlers()[0])
+		logger.SetLevel(xylog.DEBUG)
+		b.RunParallel(func(p *testing.PB) {
+			for p.Next() {
+				logger.Debug(test.GetRandomMessage())
 			}
 		})
 	})
 }
 
 func BenchmarkLoggerTextFormatter(b *testing.B) {
-	withBenchLogger(b, func(logger *xylog.Logger) {
-		var formatter = addFullMacros(xylog.NewTextFormatter())
+	test.WithBenchLogger(b, func(logger *xylog.Logger) {
+		var formatter = test.AddFullMacros(xylog.NewTextFormatter())
 		logger.Handlers()[0].SetFormatter(formatter)
 		logger.SetLevel(xylog.DEBUG)
 		b.RunParallel(func(p *testing.PB) {
 			for p.Next() {
-				logger.Debug(getRandomMessage())
+				logger.Debug(test.GetRandomMessage())
 			}
 		})
 	})
 }
 
 func BenchmarkLoggerJSONFormatter(b *testing.B) {
-	withBenchLogger(b, func(logger *xylog.Logger) {
-		var formatter = addFullMacros(xylog.NewJSONFormatter())
+	test.WithBenchLogger(b, func(logger *xylog.Logger) {
+		var formatter = test.AddFullMacros(xylog.NewJSONFormatter())
 		logger.Handlers()[0].SetFormatter(formatter)
 		logger.SetLevel(xylog.DEBUG)
 		b.RunParallel(func(p *testing.PB) {
 			for p.Next() {
-				logger.Debug(getRandomMessage())
+				logger.Debug(test.GetRandomMessage())
 			}
 		})
 	})