+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 -   Public some functions, methods.
 -   Use lock more exactly.
 -   Refactor unittests.
--   Improve performance by using buffer.
+-   Improve performance by using buffer, pool, and reducing unnecessary works.
 -   Remove the old TextFormatter. StructuredFormatter is renamed to
     TextFormatter.
 -   EventLogger is adapted with the formatter. EventLogger.JSON is removed.
+-   Remove `message` macro.
 
 # v0.1.0 (Sep 04, 2022)